- change VGMRgn::volume to m_attenDb, add set/add methods which accept either db or %amp values - update SynthRgn to have an attenDb member - add initialAttenuation generator to SF2 conversion. set SynthRgn::attenDb here rather than setting attenuation on the sample used by the region (multiple regions can use the same sample!) - remove convertLogScaleValToAtten. replaces all usages with convertPercentAmplitudeToAttenDB - Fix convertDBAttenuationToStdMidiVal(); - rename convert7bitPercentVolValToStdMidiVal to convert7bitPercentAmpValToStdMidiVal - allow rgn->sampOffset to find a matching sample by simply comparing against a samples dwOffset rather than the dwOffset - the sampColl offset - the sample data offset. - rewrite linearAmpDecayTimeToLinDBDecayTime. The new method uses two different algorithms, one for short decays and one for long

+++ src/main/formats/Akao/AkaoInstr.cpp
@@ -229,7 +229,7 @@ bool AkaoDrumKit::loadInstr() {
       addRgn(rgn);
       rgn->drumRelUnityKey = readByte(rgn_offset + 1);
       const uint16_t raw_volume = getWord(rgn_offset + 2);
-      rgn->setVolume(raw_volume / static_cast<double>(127 * 128));
+      rgn->setVolume(raw_volume / (127 * 128.0));
       rgn->addGeneralItem(rgn_offset, 1, "Associated Articulation ID");
       rgn->addGeneralItem(rgn_offset + 1, 1, "Relative Unity Key");
       rgn->addGeneralItem(rgn_offset + 2, 2, "Attenuation");
